changed the file structure for the tests
